+++ scorched/connection.py
@@ -52,6 +52,8 @@ class SolrConnection(object):
         self.update_url = self.url + "update/json"
         self.select_url = self.url + "select/"
         self.mlt_url = self.url + "mlt/"
+        self.get_url = self.url + "get/"
+        self.suggest_url = self.url + "suggest/"
         self.retry_timeout = retry_timeout
         self.max_length_get_url = max_length_get_url
         self.search_timeout = search_timeout
@@ -75,10 +77,38 @@ class SolrConnection(object):
             time.sleep(self.retry_timeout)
             return self.http_connection.request(*args, **kwargs)
 
+    def get(self, ids, fl=None):
+        """
+        Perform a RealTime Get
+        """
+        # We always send the ids parameter to force the standart output format,
+        # but use the id parameter for our actual data as `ids` can no handle
+        # ids with commas
+        params = [
+            ("ids", ""),
+            ("wt", "json"),
+        ]
+        if is_iter(ids):
+            for id in ids:
+                params.append(("id", id))
+        else:
+            params.append(("id", ids))
+        if fl:
+            params.append(("fl", ",".join(fl)))
+
+        qs = scorched.compat.urlencode(params)
+        url = "%s?%s" % (self.get_url, qs)
+
+        response = self.request("GET", url)
+        if response.status_code != 200:
+            raise scorched.exc.SolrError(response)
+        return response.text
+
     def update(self, update_doc, **kwargs):
         """
         :param update_doc: data send to solr
         :type update_doc: json data
+        :returns: json -- json string
 
         Send json to solr
         """
@@ -93,6 +123,7 @@ class SolrConnection(object):
         response = self.request('POST', url, data=body, headers=headers)
         if response.status_code != 200:
             raise scorched.exc.SolrError(response)
+        return response.text
 
     def url_for_update(self, commit=None, commitWithin=None, softCommit=None,
                        optimize=None, waitSearcher=None, expungeDeletes=None,
@@ -196,8 +227,22 @@ class SolrConnection(object):
             raise scorched.exc.SolrError(response)
         return response.text
 
+    def suggest(self, params):
+        qs = scorched.compat.urlencode(params)
+        url = "%s?%s" % (self.suggest_url, qs)
+
+        response = self.request("GET", url)
+        if response.status_code != 200:
+            raise scorched.exc.SolrError(response)
+        return response.text
+
     def mlt(self, params, content=None):
         """
+        :param params: LuceneQuery converted to a dictionary with search
+                       queries
+        :type params: dict
+        :returns: json -- json string
+
         Perform a MoreLikeThis query using the content specified
         There may be no content if stream.url is specified in the params.
         """
@@ -253,8 +298,7 @@ class SolrInterface(object):
         self.conn = SolrConnection(
             url, http_connection, mode, retry_timeout, max_length_get_url)
         self.schema = self.init_schema()
-        # we need tuples for endswith
-        self._datefields = tuple(self._extract_datefields(self.schema))
+        self._datefields = self._extract_datefields(self.schema)
 
     def init_schema(self):
         response = self.conn.request(
@@ -268,13 +312,10 @@ class SolrInterface(object):
         return response.json()['schema']
 
     def _extract_datefields(self, schema):
-        ret = [x for x in
+        ret = [x['name'] for x in
                schema['fields'] if x['type'] == 'date']
-        ret.extend([x for x in schema['dynamicFields']
+        ret.extend([x['name'] for x in schema['dynamicFields']
                     if x['type'] == 'date'])
-        if ret:
-            ret = [x['name'] for x in ret]
-            ret = [x.replace('*', '') for x in ret]
         return ret
 
     def _prepare_docs(self, docs):
@@ -286,9 +327,7 @@ class SolrInterface(object):
                 # fields
                 if value is None:
                     continue
-                if name in self._datefields:
-                    value = str(scorched.dates.solr_date(value))
-                elif name.endswith(self._datefields):
+                if scorched.dates.is_datetime_field(name, self._datefields):
                     value = str(scorched.dates.solr_date(value))
                 new_doc[name] = value
             prepared_docs.append(new_doc)
@@ -303,6 +342,7 @@ class SolrInterface(object):
         :type chunk: int
         :param kwargs: optinal -- additional arguments
         :type kwargs: dict
+        :returns: list of SolrUpdateResponse  -- A solr response object.
 
         Add a document or a list of document to solr.
         """
@@ -310,29 +350,38 @@ class SolrInterface(object):
             docs = [docs]
         # to avoid making messages too large, we break the message every
         # chunk docs.
+        ret = []
         for doc_chunk in grouper(docs, chunk):
             update_message = json.dumps(self._prepare_docs(doc_chunk))
-            self.conn.update(update_message, **kwargs)
+            ret.append(scorched.response.SolrUpdateResponse.from_json(
+                self.conn.update(update_message, **kwargs)))
+        return ret
 
     def delete_by_query(self, query, **kwargs):
         """
         :param query: criteria how witch entries should be deleted
         :type query: LuceneQuery
+        :returns: SolrUpdateResponse  -- A solr response object.
 
         Delete entries by a given query
         """
         delete_message = json.dumps({"delete": {"query": str(query)}})
-        self.conn.update(delete_message, **kwargs)
+        ret = scorched.response.SolrUpdateResponse.from_json(
+            self.conn.update(delete_message, **kwargs))
+        return ret
 
     def delete_by_ids(self, ids, **kwargs):
         """
         :param ids: ids of entries that should be deleted
         :type ids: list
+        :returns: SolrUpdateResponse  -- A solr response object.
 
         Delete entries by a given id
         """
         delete_message = json.dumps({"delete": ids})
-        self.conn.update(delete_message, **kwargs)
+        ret = scorched.response.SolrUpdateResponse.from_json(
+            self.conn.update(delete_message, **kwargs))
+        return ret
 
     def commit(self, waitSearcher=None, expungeDeletes=None, softCommit=None):
         """
@@ -346,13 +395,16 @@ class SolrInterface(object):
                            refresh the 'view' of the index in a more performant
                            manner, but without "on-disk" guarantees.
         :type softCommit: bool
+        :returns: SolrUpdateResponse  -- A solr response object.
 
         A commit operation makes index changes visible to new search requests.
         """
-        self.conn.update('{"commit": {}}', commit=True,
-                         waitSearcher=waitSearcher,
-                         expungeDeletes=expungeDeletes,
-                         softCommit=softCommit)
+        ret = scorched.response.SolrUpdateResponse.from_json(
+            self.conn.update('{"commit": {}}', commit=True,
+                             waitSearcher=waitSearcher,
+                             expungeDeletes=expungeDeletes,
+                             softCommit=softCommit))
+        return ret
 
     def optimize(self, waitSearcher=None, maxSegments=None):
         """
@@ -363,25 +415,76 @@ class SolrInterface(object):
         :param maxSegments: optional -- optimizes down to at most this number
                             of segments
         :type maxSegments: int
+        :returns: SolrUpdateResponse  -- A solr response object.
 
         An optimize is like a hard commit except that it forces all of the
         index segments to be merged into a single segment first.
         """
-        self.conn.update('{"optimize": {}}', optimize=True,
-                         waitSearcher=waitSearcher, maxSegments=maxSegments)
+        ret = scorched.response.SolrUpdateResponse.from_json(
+            self.conn.update('{"optimize": {}}', optimize=True,
+                             waitSearcher=waitSearcher,
+                             maxSegments=maxSegments))
+        return ret
 
     def rollback(self):
         """
+        :returns: SolrUpdateResponse  -- A solr response object.
+
         The rollback command rollbacks all add/deletes made to the index since
         the last commit
         """
-        self.conn.update('{"rollback": {}}')
+        ret = scorched.response.SolrUpdateResponse.from_json(
+            self.conn.update('{"rollback": {}}'))
+        return ret
 
     def delete_all(self):
         """
+        :returns: SolrUpdateResponse  -- A solr response object.
+
         Delete everything
         """
-        self.delete_by_query(self.Q(**{"*": "*"}))
+        return self.delete_by_query(self.Q(**{"*": "*"}))
+
+    def get(self, ids, fields=None):
+        """
+        RealTime Get document(s) by id(s)
+
+        :param ids: id(s) of the document(s)
+        :type ids: list, string or int
+        :param fields: optional -- list of fields to return
+        :type fileds: list of strings
+        """
+        ret = scorched.response.SolrResponse.from_get_json(
+            self.conn.get(ids, fields), self._datefields)
+        return ret
+
+    def suggest(self, q=None, **kwargs):
+        boolArgs = ['reload', 'reloadAll', 'build', 'buildAll']
+        otherArgs = ['count', 'dictionary']
+        params = {
+            "suggest": "true",
+            "wt": "json",
+        }
+        if q:
+            params["q"] = q
+        for arg, v in kwargs.items():
+            if arg not in boolArgs + otherArgs:
+                raise TypeError(
+                    "{} is not a valid argument for suggest()".format(arg))
+            if arg in boolArgs:
+                params["suggest." + arg] = "true" if v else "false"
+            else:
+                params["suggest." + arg] = v
+
+        data = json.loads(self.conn.suggest(params))
+
+        if "suggest" in data:
+            suggester = data["suggest"].popitem()[1]
+            suggest_term = suggester.popitem()[1]
+            suggestions = suggest_term["suggestions"]
+            return suggestions
+        else:
+            return None
 
     def search(self, **kwargs):
         """
@@ -408,6 +511,8 @@ class SolrInterface(object):
 
     def mlt_search(self, content=None, **kwargs):
         """
+        :returns: SolrResponse  -- A solr response object.
+
         Mlt search solr
         """
         params = scorched.search.params_from_dict(**kwargs)